Make String_from_float take double which simplifies llvm codegen.

+++ wlib/src/wich.c
@@ -183,7 +183,7 @@ String *String_from_int(int value) {
 	return String_new(s);
 }
 
-String *String_from_float(float value) {
+String *String_from_float(double value) {
 	char *s = calloc(20, sizeof(char));
 	char buf[50];
 	sprintf(buf,"%1.2f",value);
